refactor: normalize naming of commands
