+++ README.md
@@ -61,8 +61,8 @@ This package ships with the following commands;
 |-----------------------|------------------------------------|------------------------------------------------|-------------------------------------------------------------------------------------------------------|
 | `plugins:meta`        | None                               | [--update-all or -u] [--plugins]               | Downloads metadata for plugins.                                                                       |
 | `themes:meta`         | None                               | [--update-all or -u] [--themes]                | Downloads metadata for themes.                                                                        |
-| `download:plugins`    | [num-to-pull=latest]               | [--plugins [--force-download or -f]            | Downloads any plugins outstanding from the last time the command was run. Defaults to latest version, |
-| `download:themes`     | [num-to-pull=latest]               | [--themes] [--force-download or  -f]           | Downloads any themes outstanding from the last time the command was run. Defaults to latest version.  |
+| `plugins:download`    | [num-to-pull=latest]               | [--plugins [--force-download or -f]            | Downloads any plugins outstanding from the last time the command was run. Defaults to latest version, |
+| `themes:download`     | [num-to-pull=latest]               | [--themes] [--force-download or  -f]           | Downloads any themes outstanding from the last time the command was run. Defaults to latest version.  |
 | `util:upload`         | action<plugins, themes>            | [--slugs] [--limit] [--offset] [--clean or -c] | Uploads any downloaded plugins/themes to your file system (right now supports S3).                    |
 | `util:clean`          | None                               | None                                           | Cleans the data directory of any files that remain.                                                   |  
 | `run:all`             | [asset-type<all, plugins, themes>] | None                                           | Runs the four commands for themes/plugins or both. A shortcut to a full run of the downloader.        |
