+++ bin/aspiresync
@@ -1,13 +1,13 @@
 #!/usr/bin/env php
 <?php
 
-use AspirePress\AspireSync\Commands\Plugins\DownloadPluginsCommand;
-use AspirePress\AspireSync\Commands\Plugins\DownloadPluginsPartialCommand;
+use AspirePress\AspireSync\Commands\Plugins\PluginsDownloadCommand;
+use AspirePress\AspireSync\Commands\Plugins\PluginsDownloadPartialCommand;
 use AspirePress\AspireSync\Commands\Plugins\InternalPluginDownloadCommand;
 use AspirePress\AspireSync\Commands\Plugins\PluginsMetaCommand;
 use AspirePress\AspireSync\Commands\RunAllCommand;
-use AspirePress\AspireSync\Commands\Themes\DownloadThemesCommand;
-use AspirePress\AspireSync\Commands\Themes\DownloadThemesPartialCommand;
+use AspirePress\AspireSync\Commands\Themes\ThemesDownloadCommand;
+use AspirePress\AspireSync\Commands\Themes\ThemesDownloadPartialCommand;
 use AspirePress\AspireSync\Commands\Themes\InternalThemeDownloadCommand;
 use AspirePress\AspireSync\Commands\Themes\ThemesMetaCommand;
 use Symfony\Component\Config\FileLocator;
@@ -34,12 +34,12 @@ $APP = new Application();
 
 $APP->add($CONTAINER->get(RunAllCommand::class));
 $APP->add($CONTAINER->get(InternalPluginDownloadCommand::class));
-$APP->add($CONTAINER->get(DownloadPluginsCommand::class));
+$APP->add($CONTAINER->get(PluginsDownloadCommand::class));
 $APP->add($CONTAINER->get(PluginsMetaCommand::class));
-$APP->add($CONTAINER->get(DownloadPluginsPartialCommand::class));
+$APP->add($CONTAINER->get(PluginsDownloadPartialCommand::class));
 $APP->add($CONTAINER->get(ThemesMetaCommand::class));
-$APP->add($CONTAINER->get(DownloadThemesCommand::class));
-$APP->add($CONTAINER->get(DownloadThemesPartialCommand::class));
+$APP->add($CONTAINER->get(ThemesDownloadCommand::class));
+$APP->add($CONTAINER->get(ThemesDownloadPartialCommand::class));
 $APP->add($CONTAINER->get(InternalThemeDownloadCommand::class));
 // $application->add($container->get(UtilUploadCommand::class));    // borked
