+++ src/Commands/Themes/ThemesDownloadCommand.php
@@ -10,6 +10,7 @@ use AspirePress\AspireSync\Services\Themes\ThemeListService;
 use AspirePress\AspireSync\Services\Themes\ThemesMetadataService;
 use AspirePress\AspireSync\Utilities\GetItemsFromSourceTrait;
 use AspirePress\AspireSync\Utilities\ProcessWaitUtil;
+use AspirePress\AspireSync\Utilities\StringUtil;
 use AspirePress\AspireSync\Utilities\VersionUtil;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
@@ -18,66 +19,58 @@ use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Process\Process;
 
-class DownloadThemesPartialCommand extends AbstractBaseCommand
+class ThemesDownloadCommand extends AbstractBaseCommand
 {
     use GetItemsFromSourceTrait;
 
-    public function __construct(private ThemeListService $themeListService, private ThemesMetadataService $themesMetadataService, private StatsMetadataService $statsMetadataService)
+    public function __construct(private ThemeListService $themeListService, private ThemesMetadataService $themeMetadataService, private StatsMetadataService $statsMetadataService)
     {
         parent::__construct();
     }
 
     protected function configure(): void
     {
-        $this->setName('download:themes:partial')
-            ->setAliases(['themes:partial'])
-            ->setDescription('Pulls a partial number of themes based on the full list of themes')
-            ->addArgument('num-to-pull', InputArgument::REQUIRED, 'Number of themes to pull')
-            ->addArgument('offset', InputArgument::OPTIONAL, 'Offset to start pulling from', 0)
-            ->AddOption('versions', null, InputOption::VALUE_OPTIONAL, 'Number of versions to request', 'latest')
+        $this->setName('themes:download')
+            ->setDescription('Grabs themes (with number of specified versions or explicitly specified themes) from the origin repo')
+            ->addArgument('num-versions', InputArgument::OPTIONAL, 'Number of versions to request', 'latest')
+            ->addOption('themes', null, InputOption::VALUE_OPTIONAL, 'List of themes to request')
             ->addOption('force-download', 'f', InputOption::VALUE_NONE, 'Force download even if file exists')
-            ->addOption('download-all', 'd', InputOption::VALUE_NONE, 'Download all themes (limited by offset and limit)');
+            ->addOption('download-all', 'd', InputOption::VALUE_NONE, 'Download all themes');
     }
 
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
-        $this->always("Running command {$this->getName()}");
+        $this->always('Running command ' . $this->getName());
         $this->startTimer();
-        $numVersions = $input->getOption('versions');
-        $numToPull   = (int) $input->getArgument('num-to-pull');
-        $offset      = (int) $input->getArgument('offset');
+        $numVersions = $input->getArgument('num-versions');
+        $themesList  = $input->getOption('themes');
 
-        $this->debug('Getting list of themes...');
-        if ($input->getOption('download-all')) {
-            $themesToUpdate = $this->themeListService->getUpdatedListOfItems([], 'default');
-        } else {
-            $themesToUpdate = $this->themeListService->getUpdatedListOfItems([]);
+        if ($themesList) {
+            $themesList = StringUtil::explodeAndTrim($themesList);
         }
 
-        $totalthemes = count($themesToUpdate);
+        $this->info('Getting list of themes...');
+        if ($input->getOption('download-all')) {
+            $themesToUpdate = $this->themeListService->getUpdatedListOfItems($themesList, 'default');
+        } else {
+            $themesToUpdate = $this->themeListService->getUpdatedListOfItems($themesList);
+        }
 
-        $this->debug($totalthemes . ' themes to download...');
-
-        if ($totalthemes === 0) {
-            $this->success('No themes to download...exiting...');
+        $this->info(count($themesToUpdate) . ' themes to download...');
+        if (count($themesToUpdate) === 0) {
+            $this->always('No themes to download...exiting...');
             return Command::SUCCESS;
         }
-
-        if ($offset > $totalthemes) {
-            $this->failure('Offset is greater than total themes...exiting...');
-            return Command::FAILURE;
-        }
-        $this->info('Limiting theme download to ' . $numToPull . ' themes... (offset by ' . $offset . ')');
-        $themesToUpdate = array_slice($themesToUpdate, $offset, $numToPull);
 
         $processes = [];
 
         foreach ($themesToUpdate as $theme => $versions) {
-            $versions    = $this->determineVersionsToDownload($theme, $versions, $numVersions);
+            $versions = $this->determineVersionsToDownload($theme, $versions, $numVersions);
+
             $versionList = implode(',', $versions);
 
             if (empty($versionList)) {
-                $this->info('No versions found for ' . $theme . '...skipping...');
+                $this->notice('No downloadable versions found for ' . $theme . '...skipping...');
                 continue;
             }
 
@@ -95,18 +88,18 @@ class DownloadThemesPartialCommand extends AbstractBaseCommand
 
             $process = new Process($command);
             $process->start();
-            $processes[] = $process;
 
+            $processes[] = $process;
             if (count($processes) >= 24) {
                 $this->debug('Max processes reached...waiting for space...');
                 $stats = ProcessWaitUtil::wait($processes);
-                $this->processStats($stats);
                 $this->info($stats);
+                $this->processStats($stats);
                 $this->debug('Process ended; starting another...');
             }
         }
 
-        $this->debug('Waiting for all processes to finish...');
+        $this->info('Waiting for all processes to finish...');
 
         $stats = ProcessWaitUtil::waitAtEndOfScript($processes);
         foreach ($stats as $stat) {
@@ -114,10 +107,10 @@ class DownloadThemesPartialCommand extends AbstractBaseCommand
             $this->info($stat);
         }
 
-        $this->debug('All processes finished!');
+        $this->info('All processes finished...');
 
+        // Output statistics
         $this->endTimer();
-
         $this->always($this->getRunInfo($this->getCalculatedStats()));
         $this->statsMetadataService->logStats($this->getName(), $this->stats);
         return Command::SUCCESS;
@@ -134,6 +127,6 @@ class DownloadThemesPartialCommand extends AbstractBaseCommand
             'latest' => [VersionUtil::getLatestVersion($versions)],
             default => VersionUtil::limitVersions(VersionUtil::sortVersions($versions), (int) $numToDownload),
         };
-        return $this->themesMetadataService->getUnprocessedVersions($theme, $download);
+        return $this->themeMetadataService->getUnprocessedVersions($theme, $download);
     }
 }
