+++ src/Commands/Themes/ThemesDownloadPartialCommand.php
@@ -2,12 +2,12 @@
 
 declare(strict_types=1);
 
-namespace AspirePress\AspireSync\Commands\Plugins;
+namespace AspirePress\AspireSync\Commands\Themes;
 
 use AspirePress\AspireSync\Commands\AbstractBaseCommand;
-use AspirePress\AspireSync\Services\Plugins\PluginListService;
-use AspirePress\AspireSync\Services\Plugins\PluginMetadataService;
 use AspirePress\AspireSync\Services\StatsMetadataService;
+use AspirePress\AspireSync\Services\Themes\ThemeListService;
+use AspirePress\AspireSync\Services\Themes\ThemesMetadataService;
 use AspirePress\AspireSync\Utilities\GetItemsFromSourceTrait;
 use AspirePress\AspireSync\Utilities\ProcessWaitUtil;
 use AspirePress\AspireSync\Utilities\VersionUtil;
@@ -18,74 +18,72 @@ use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Process\Process;
 
-class DownloadPluginsPartialCommand extends AbstractBaseCommand
+class ThemesDownloadPartialCommand extends AbstractBaseCommand
 {
     use GetItemsFromSourceTrait;
 
-    public function __construct(private PluginListService $pluginListService, private PluginMetadataService $pluginMetadataService, private StatsMetadataService $statsMetadataService)
+    public function __construct(private ThemeListService $themeListService, private ThemesMetadataService $themesMetadataService, private StatsMetadataService $statsMetadataService)
     {
         parent::__construct();
     }
 
     protected function configure(): void
     {
-        $this->setName('download:plugins:partial')
-            ->setAliases(['plugins:partial'])
-            ->setDescription('Pulls a partial number of plugins based on the full list of plugins')
-            ->addArgument('num-to-pull', InputArgument::REQUIRED, 'Number of plugins to pull')
+        $this->setName('themes:download:partial')
+            ->setDescription('Pulls a partial number of themes based on the full list of themes')
+            ->addArgument('num-to-pull', InputArgument::REQUIRED, 'Number of themes to pull')
             ->addArgument('offset', InputArgument::OPTIONAL, 'Offset to start pulling from', 0)
             ->AddOption('versions', null, InputOption::VALUE_OPTIONAL, 'Number of versions to request', 'latest')
             ->addOption('force-download', 'f', InputOption::VALUE_NONE, 'Force download even if file exists')
-            ->addOption('download-all', 'd', InputOption::VALUE_NONE, 'Download all plugins (limited by offset and limit)');
+            ->addOption('download-all', 'd', InputOption::VALUE_NONE, 'Download all themes (limited by offset and limit)');
     }
 
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
         $this->always("Running command {$this->getName()}");
-
         $this->startTimer();
         $numVersions = $input->getOption('versions');
         $numToPull   = (int) $input->getArgument('num-to-pull');
         $offset      = (int) $input->getArgument('offset');
 
-        $this->debug('Getting list of plugins...');
+        $this->debug('Getting list of themes...');
         if ($input->getOption('download-all')) {
-            $pluginsToUpdate = $this->pluginListService->getUpdatedListOfItems([], 'default');
+            $themesToUpdate = $this->themeListService->getUpdatedListOfItems([], 'default');
         } else {
-            $pluginsToUpdate = $this->pluginListService->getUpdatedListOfItems([]);
+            $themesToUpdate = $this->themeListService->getUpdatedListOfItems([]);
         }
 
-        $totalPlugins = count($pluginsToUpdate);
+        $totalthemes = count($themesToUpdate);
 
-        $this->debug($totalPlugins . ' plugins to download...');
+        $this->debug($totalthemes . ' themes to download...');
 
-        if ($totalPlugins === 0) {
-            $this->success('No plugins to download...exiting...');
+        if ($totalthemes === 0) {
+            $this->success('No themes to download...exiting...');
             return Command::SUCCESS;
         }
 
-        if ($offset > $totalPlugins) {
-            $this->failure('Offset is greater than total plugins...exiting...');
+        if ($offset > $totalthemes) {
+            $this->failure('Offset is greater than total themes...exiting...');
             return Command::FAILURE;
         }
-        $this->info('Limiting plugin download to ' . $numToPull . ' plugins... (offset by ' . $offset . ')');
-        $pluginsToUpdate = array_slice($pluginsToUpdate, $offset, $numToPull);
+        $this->info('Limiting theme download to ' . $numToPull . ' themes... (offset by ' . $offset . ')');
+        $themesToUpdate = array_slice($themesToUpdate, $offset, $numToPull);
 
         $processes = [];
 
-        foreach ($pluginsToUpdate as $plugin => $versions) {
-            $versions    = $this->determineVersionsToDownload($plugin, $versions, $numVersions);
+        foreach ($themesToUpdate as $theme => $versions) {
+            $versions    = $this->determineVersionsToDownload($theme, $versions, $numVersions);
             $versionList = implode(',', $versions);
 
             if (empty($versionList)) {
-                $this->notice('No versions found for ' . $plugin . '...skipping...');
+                $this->info('No versions found for ' . $theme . '...skipping...');
                 continue;
             }
 
             $command = [
                 'aspiresync',
-                'internal:plugin-download',
-                $plugin,
+                'internal:theme-download',
+                $theme,
                 $versionList,
                 $numVersions,
             ];
@@ -121,7 +119,6 @@ class DownloadPluginsPartialCommand extends AbstractBaseCommand
 
         $this->always($this->getRunInfo($this->getCalculatedStats()));
         $this->statsMetadataService->logStats($this->getName(), $this->stats);
-
         return Command::SUCCESS;
     }
 
@@ -129,13 +126,13 @@ class DownloadPluginsPartialCommand extends AbstractBaseCommand
      * @param string[] $versions
      * @return array<int, string>
      */
-    private function determineVersionsToDownload(string $plugin, array $versions, string $numToDownload): array
+    private function determineVersionsToDownload(string $theme, array $versions, string $numToDownload): array
     {
         $download = match ($numToDownload) {
             'all' => $versions,
             'latest' => [VersionUtil::getLatestVersion($versions)],
             default => VersionUtil::limitVersions(VersionUtil::sortVersions($versions), (int) $numToDownload),
         };
-        return $this->pluginMetadataService->getUnprocessedVersions($plugin, $download);
+        return $this->themesMetadataService->getUnprocessedVersions($theme, $download);
     }
 }
