+++ src/Services/Plugins/PluginDownloadFromWpService.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace AspirePress\AspireSync\Services\Plugins;
 
 use AspirePress\AspireSync\Services\Interfaces\DownloadServiceInterface;
+use Exception;
 use GuzzleHttp\Client as GuzzleClient;
 use GuzzleHttp\Exception\ClientException;
 use RuntimeException;
@@ -18,6 +19,7 @@ class PluginDownloadFromWpService implements DownloadServiceInterface
     ) {
     }
 
+    /** @return array<string,string[]> */
     public function download(string $slug, array $versions, bool $force = false): array
     {
         $downloadable = $this->pluginMetadataService->getDownloadUrlsForVersions($slug, $versions);
@@ -30,40 +32,44 @@ class PluginDownloadFromWpService implements DownloadServiceInterface
 
         $outcomes = [];
         foreach ($downloadable as $version => $url) {
-            $filePath = "/opt/aspiresync/data/plugins/$slug.$version.zip";
+            $outcome            = $this->downloadOne($url, $slug, $version, $force);
+            $outcomes[$outcome] = $version;
+        }
+        return $outcomes;
+    }
 
-            if (file_exists($filePath) && ! $force) {
-                $outcomes['304 Not Modified'][] = $version;
-                $hash                           = $this->calculateHash($filePath);
-                $this->pluginMetadataService->setVersionToDownloaded($slug, (string) $version, $hash);
-                continue;
-            }
-            try {
-                $response = $this->guzzle->request('GET', $url, ['allow_redirects' => true, 'sink' => $filePath]);
-                $outcomes["{$response->getStatusCode()} {$response->getReasonPhrase()}"][] = $version;
-                if (filesize($filePath) === 0) {
-                    unlink($filePath);
-                }
-                $hash = $this->calculateHash($filePath);
-                $this->pluginMetadataService->setVersionToDownloaded($slug, (string) $version, $hash);
-            } catch (ClientException $e) {
-                if (method_exists($e, 'getResponse')) {
-                    $response = $e->getResponse();
-                    $outcomes["{$response->getStatusCode()} {$response->getReasonPhrase()}"][] = $version;
-                    if ($response->getStatusCode() === 404) {
-                        $this->pluginMetadataService->setVersionToDownloaded($slug, (string) $version);
-                    }
-                } else {
-                    $outcomes[$e->getMessage()][] = $version;
-                }
-                unlink($filePath);
-            } catch (RuntimeException $e) {
-                $outcomes[$e->getMessage()][] = $version;
-                @unlink($filePath);
-            }
+    private function downloadOne(string $url, string $slug, string $version, bool $force = false): string
+    {
+        $filePath = "/opt/aspiresync/data/plugins/$slug.$version.zip";
+
+        if (file_exists($filePath) && ! $force) {
+            $hash = $this->calculateHash($filePath);
+            $this->pluginMetadataService->setVersionToDownloaded($slug, $version, $hash);
+            return '304 Not Modified';
         }
 
-        return $outcomes;
+        try {
+            $response = $this->guzzle->request('GET', $url, ['allow_redirects' => true, 'sink' => $filePath]);
+            if (filesize($filePath) === 0) {
+                unlink($filePath);
+            }
+            $this->pluginMetadataService->setVersionToDownloaded($slug, $version, $this->calculateHash($filePath));
+            return "{$response->getStatusCode()} {$response->getReasonPhrase()}";
+        } catch (ClientException $e) {
+            @unlink($filePath);
+            if (method_exists($e, 'getResponse')) {
+                $response = $e->getResponse();
+                if ($response->getStatusCode() === 404) {
+                    $this->pluginMetadataService->setVersionToDownloaded($slug, $version);
+                }
+                return "{$response->getStatusCode()} {$response->getReasonPhrase()}";
+            } else {
+                return $e->getMessage();
+            }
+        } catch (Exception $e) {
+            @unlink($filePath);
+            return $e->getMessage();
+        }
     }
 
     private function calculateHash(string $filePath): string
